Refactor: Move ConfidentialMessageService.SendConfidentialMessageResult one level up
